Update type of cycle_time_target in arguments.py to Int from Float.
[story=PO-284 story_url=https://urjuna.atlassian.net/browse/PO-284]

+++ polaris/analytics/service/graphql/arguments.py
@@ -137,7 +137,7 @@ class ResponseTimeConfidenceTrendsParameters(AggregateMetricsTrendsParameters):
         required=True,
         description="Target lead time in days for which confidence is measured"
     )
-    cycle_time_target = graphene.Float(
+    cycle_time_target = graphene.Int(
         required=True,
         description="Target cycle time in days for which confidence is measured"
     )
